pushed to master changes from freebsd-buildserver branch

+++ scripts/pkginstall.sh
@@ -15,12 +15,8 @@ if [ -z "${LOGFILE:-}" ]; then
   exit 1
 fi
 
-PKGFILE=${PKGFILE:-${LOCALDIR}/conf/${PACK_PROFILE}-packages};
-
-#if [ ! -f ${PKGFILE} ]; then
- # return
-#fi
-touch ${PKGFILE}
+pkgfile=${pkgfile:-${LOCALDIR}/conf/${PACK_PROFILE}-packages};
+touch ${pkgfile}
 
 # Search main file package for include dependecies
 # and build an depends file ( depends )
@@ -53,8 +49,8 @@ for left_files in ports ghostbsd pcbsd gbi ; do
     rm -Rf ${BASEDIR}/${left_files}
 done
 
-if [ -f ${BASEDIR}/usr/local/etc/repos/GhostBSD.conf ]; then
-    rm -f  ${BASEDIR}/usr/local/etc/repos/GhostBSD.conf
+if [ -f ${BASEDIR}/usr/local/etc/repos/DesktopBSD.conf ]; then
+    rm -f  ${BASEDIR}/usr/local/etc/repos/DesktopBSD.conf
 fi
 
 #mounts ${BASEDIR}/var/run because it's needed when building ports in chroot
@@ -66,25 +62,55 @@ fi
 cp /etc/resolv.conf ${BASEDIR}/etc/resolv.conf
 
 PLOGFILE=".log_pkginstall"
-echo "Installing packages listed in ${PKGFILE}"
+echo "Installing packages listed in ${pkgfile}"
 
 # cp resolv.conf for fetching packages
-cp $PKGFILE ${BASEDIR}/mnt
+cp $pkgfile ${BASEDIR}/mnt
 
 sed -i '' 's@signature_type: "fingerprints"@#signature_type: "fingerprints"@g' ${BASEDIR}/etc/pkg/FreeBSD.conf
 
+if ! ${USE_DISTROREPO} ; then
+if [ ! -d ${BASEDIR}/dist/ports ]; then
+
 # prepares ports file backend an mounts it over /dist/ports
-PSIZE=$(echo "${PORTS_SIZE}*1024^2" | bc | cut -d . -f1)
-dd if=/dev/zero of=${BASEDIR}/ports.ufs bs=1k count=1 seek=$((${PSIZE} - 1))
-PDEVICE=$(mdconfig -a -t vnode -f ${BASEDIR}/ports.ufs)
-echo $PDEVICE >${BASEDIR}/pdevice
-newfs -o space /dev/$PDEVICE
-mkdir -p ${BASEDIR}/dist/ports
-mount -o noatime /dev/$PDEVICE  ${BASEDIR}/dist/ports
+    PSIZE=$(echo "${PORTS_SIZE}*1024^2" | bc | cut -d . -f1)
+    dd if=/dev/zero of=${BASEDIR}/ports.ufs bs=1k count=1 seek=$((${PSIZE} - 1))
+    PDEVICE=$(mdconfig -o async -o cluster -S 4096 -a -t vnode -f ${BASEDIR}/ports.ufs)
+    echo $PDEVICE >${BASEDIR}/pdevice
+    newfs -o space /dev/$PDEVICE
+    mkdir -p ${BASEDIR}/dist/ports
+    mount -o noatime /dev/$PDEVICE  ${BASEDIR}/dist/ports
 
 # prepares ports tree
-portsnap fetch
-portsnap extract -p ${BASEDIR}/dist/ports
+    portsnap fetch
+    portsnap extract -p ${BASEDIR}/dist/ports
+fi
+
+buildpkg()
+{
+# prepares buildpkg.sh script to add packages under chroot
+cat > ${BASEDIR}/mnt/buildpkg.sh << "EOF"
+#!/bin/sh 
+
+FORCE_PKG_REGISTER=true
+export FORCE_PKG_REGISTER
+ln -sf /dist/ports /usr/ports
+
+# builds pkg from ports to avoid Y/N question
+cd /usr/ports/ports-mgmt/pkg
+make
+make install
+
+# pkg install part
+cd /mnt
+rm buildpkg.sh
+EOF
+
+# run addpkg.sh in chroot to add packages
+chrootcmd="chroot ${BASEDIR} sh /mnt/buildpkg.sh"
+$chrootcmd
+}
+fi
 
 # prepares addpkg.sh script to add packages under chroot
 cat > ${BASEDIR}/mnt/addpkg.sh << "EOF"
@@ -92,7 +118,6 @@ cat > ${BASEDIR}/mnt/addpkg.sh << "EOF"
 
 FORCE_PKG_REGISTER=true
 export FORCE_PKG_REGISTER
-
 ln -sf /dist/ports /usr/ports
 
 # builds pkg from ports to avoid Y/N question
@@ -112,13 +137,20 @@ pkgaddcmd="pkg install -y "
 while read pkgc; do
     if [ -n "${pkgc}" ] ; then
     echo "Installing package $pkgc"
+    #echo "fetching package $pkgc"
+    #pkg fetch -y -d $pkgc
     echo "Running $pkgaddcmd ${pkgc}" >> ${PLOGFILE} 2>&1
     $pkgaddcmd $pkgc >> ${PLOGFILE} 2>&1
+    if [ $? -ne 0 ] ; then
+        echo "$pkgc not found in repos" >> ${PLOGFILE} 2>&1
+        exit 1
+    fi
     fi
 done < $pkgfile
 
 rm addpkg.sh
 rm $pkgfile
+#pkg check -y -a
 EOF
 
 # run addpkg.sh in chroot to add packages
